pyproject.toml
[tool.poetry]
name = "poc-rag"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
ipykernel = "^6.29.5"
langchain-openai = "^0.2.14"
langchain-community = "^0.3.14"
pandas = "^1.4.4"
streamlit = "^1.13.0"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pinecone = {extras = ["grpc"], version = "^5.4.2"}
easyocr = "^1.7.2"
pymupdf = "^1.25.1"
ratelimit = "^2.2.1"
pydantic = "^2.10.5"
pydantic-settings = "^2.7.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
